# Performance/safety considerations for `partition`